+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support for `nomic-ai/nomic-embed-text-v1`
 - Added support for `intfloat/e5-base-v2` to `Model`
+- Added support for `BAAI/bge-base-en-v1.5` to `Model`
 
 ## 1.0.1 (2024-08-27)